snakemake: Refactor all rules to remove duplicate code.
This commit refactors the Snakefile and related scripts to remove
duplicate code and to simplify the workflows of the pipeline.

As many config names are changed and added in the config.yaml file
and new functions are adopted, there is no compatibility with
previous versions.

In this implementation, the list of the target repositories are
no longer downloaded automatically for Snakemake and Nextflow.
The list to clone now should be provided in the config.yaml file.
This change is to provide more flexibility in the choice of
repositories to analyze.

CHANGED FILES

* modified: workflow/Snakefile
* modified: workflow/config.yaml.example
* modified: workflow/requirements.txt
* renamed:  workflow/scripts/analysis_git_repository.py
            -> workflow/scripts/git_analyzer.py
* modified: workflow/scripts/git_providers.py
* modified: workflow/scripts/storage.py